demo/requirements.txt
numpy
scipy
ffmpeg-python
cv2
matplotlib
tensorflow
